resolved review isssues
