--- src/main/java/Helpers/User.java
@@ -1,11 +1,0 @@
-package Helpers;
-import lombok.*;
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class User {
-
-    String email;
-    String password;
-    String name;
-}
