+++ src/main/java/helpers/User.java
@@ -1,0 +1,11 @@
+package helpers;
+import lombok.*;
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class User {
+
+    String email;
+    String password;
+    String name;
+}
