+++ src/test/java/site/nomoreparties/stellarburgers/BaseTest.java
@@ -5,19 +5,26 @@ import org.junit.After;
 import org.junit.Before;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.firefox.FirefoxDriver;
 import site.nomoreparties.stellarburgers.pageobject.page.MainPage;
 
-import java.time.Duration;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
 public class BaseTest {
     protected WebDriver driver;
-
+    Properties properties = new Properties();
     @Before
     public void setUpDriver () {
-driver=getDriver("yandex");
+
+        try (FileInputStream fileInputStream = new FileInputStream("src/test/resources/config.properties")) {
+            properties.load(fileInputStream);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        driver=getDriver(properties.getProperty("BROWSER"));
 driver.get(MainPage.URL);
 driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
 
@@ -32,15 +39,15 @@ private WebDriver getDriver (String driverType) {
             case "chrome":
                 WebDriverManager.chromedriver().setup();
                 return new ChromeDriver();
-            case "firefox":
+          /*  case "firefox":
                 WebDriverManager.firefoxdriver().clearDriverCache().setup();
                 System.setProperty("webdriver.firefox.bin", "/Applications/Firefox.app/Contents/MacOS/firefox");
-                return new FirefoxDriver();
+                return new FirefoxDriver(); */
             case "yandex":
-                System.setProperty("webdriver.chrome.driver","src/main/resources/chromedriver");
-                ChromeOptions options = new ChromeOptions();
-                options.setBinary("/Applications/Yandex.app/Contents/MacOS/Yandex");
-                return new ChromeDriver(options);
+                System.setProperty("webdriver.chrome.driver","src/test/resources/yandexdriver");
+           //     ChromeOptions options = new ChromeOptions();
+            //    options.setBinary(properties.getProperty("ya.binary"));
+                    return new ChromeDriver();
             default:
                 throw new IllegalArgumentException("такой driverType не поддерживается ");
         }
