+++ src/test/java/site/nomoreparties/stellarburgers/ConstructorTest.java
@@ -1,19 +1,9 @@
 package site.nomoreparties.stellarburgers;
 
-import Helpers.UserClient;
-import com.github.javafaker.Faker;
 import io.qameta.allure.junit4.DisplayName;
-import io.restassured.response.ValidatableResponse;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.JavascriptExecutor;
-import site.nomoreparties.stellarburgers.pageobject.fragment.HeaderFragment;
-import site.nomoreparties.stellarburgers.pageobject.page.AccountPage;
-import site.nomoreparties.stellarburgers.pageobject.page.LoginPage;
 import site.nomoreparties.stellarburgers.pageobject.page.MainPage;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
