+++ src/test/resources/config.properties
@@ -1,0 +1,1 @@
+BROWSER=yandex
